+++ CHANGELOG.md
@@ -559,3 +559,5 @@ All notable changes to Project Genesis will be documented in this file.
 
 - Implemented `is_lava_dweller` trait. Entities with this trait treat `lava` terrain as a shelter, gaining increased defense and energy recovery.
 - [x] Implemented `is_ice_walker` trait. Entities with this trait do not consume extra stamina from elevation changes when moving on ice terrain.
+
+- Implemented `is_fire_dancer` trait. Entities with this trait gain energy and are immune to damage when inside a fire event.